+++ src/app/routes/people.ts
@@ -208,6 +208,11 @@ peopleRouter.get<ParamsDictionary>(
     validator,
     async (req, res, next) => {
         try {
+            const accountService = new cinerino.chevre.service.Account({
+                endpoint: cinerino.credentials.chevre.endpoint,
+                auth: chevreAuthClient,
+                project: { id: req.project.id }
+            });
             const productService = new cinerino.chevre.service.Product({
                 endpoint: cinerino.credentials.chevre.endpoint,
                 auth: chevreAuthClient,
@@ -247,6 +252,7 @@ peopleRouter.get<ParamsDictionary>(
                         project: req.project,
                         conditions: searchConditions
                     })({
+                        account: accountService,
                         ownershipInfo: ownershipInfoService
                     });