+++ src/app/routes/people/me/ownershipInfos.ts
@@ -56,6 +56,11 @@ ownershipInfosRouter.get(
     validator,
     async (req, res, next) => {
         try {
+            const accountService = new cinerino.chevre.service.Account({
+                endpoint: cinerino.credentials.chevre.endpoint,
+                auth: chevreAuthClient,
+                project: { id: req.project.id }
+            });
             const productService = new cinerino.chevre.service.Product({
                 endpoint: cinerino.credentials.chevre.endpoint,
                 auth: chevreAuthClient,
@@ -93,6 +98,7 @@ ownershipInfosRouter.get(
                         project: req.project,
                         conditions: searchConditions
                     })({
+                        account: accountService,
                         ownershipInfo: ownershipInfoService
                     });