+++ src/app/routes/transactions/placeOrder.ts
@@ -438,6 +438,11 @@ export async function authorizePointAward(req: Request): Promise<cinerino.factor
     const actionRepo = new cinerino.repository.Action(mongoose.connection);
     const transactionRepo = new cinerino.repository.Transaction(mongoose.connection);
 
+    const accountService = new cinerino.chevre.service.Account({
+        endpoint: cinerino.credentials.chevre.endpoint,
+        auth: chevreAuthClient,
+        project: { id: req.project.id }
+    });
     const ownershipInfoService = new cinerino.chevre.service.OwnershipInfo({
         endpoint: cinerino.credentials.chevre.endpoint,
         auth: chevreAuthClient,
@@ -481,7 +486,7 @@ export async function authorizePointAward(req: Request): Promise<cinerino.factor
                         project: { id: req.project.id },
                         now: now,
                         accountType: membershipPointsEarnedUnitText
-                    })({ ownershipInfo: ownershipInfoService });
+                    })({ account: accountService, ownershipInfo: ownershipInfoService });
 
                     givePointAwardParams.push({
                         object: {